+++ src/test-integration/resources/test-instrumented-server.js
@@ -21,7 +21,7 @@ if (!window.jscoverage_report) {
         //console.log("Configuring async jsreport cb");
         request.onreadystatechange = function () {
             //console.log("jsreport cb called");
-            if (request.readyState == 4 && callback) {
+            if (request.readyState === 4 && callback) {
                 callback(request);
             }
         };
@@ -231,7 +231,7 @@ function jscoverage_pad(s) {
 }
 
 function jscoverage_html_escape(s) {
-    return s.replace(/[<>\&\"\']/g, function (c) {
+    return s.replace(/[<>&"']/g, function (c) {
         return '&#' + c.charCodeAt(0) + ';';
     });
 }
@@ -268,13 +268,14 @@ if (!jsCover_isolateBrowser) {
 
     try {
         if (typeof top === 'object' && top !== null && top._$jscoverage) {
-            this._$jscoverage = top._$jscoverage;
+            _$jscoverage = top._$jscoverage;
         }
     } catch (e) {
     }
 }
-if (!this._$jscoverage) {
-    this._$jscoverage = {};
+var _$jscoverage;
+if (!_$jscoverage) {
+    _$jscoverage = {};
 }
 if (! _$jscoverage['test-simple.js']) {
   _$jscoverage['test-simple.js'] = {};